requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
django-cors-headers==4.2.0
black==24.8.0
click==8.1.7
django-cors-headers==4.2.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.47
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
